[feat] : 서재 기본 레이아웃 구현
