--- src/pages/Library.tsx
@@ -1,35 +1,0 @@
-import { BsThreeDotsVertical } from 'react-icons/bs';
-import { HiPlus } from 'react-icons/hi';
-import { useNavigate } from 'react-router-dom';
-
-import Header from 'components/ui/Header';
-
-const Library = () => {
-  const navigate = useNavigate();
-  const handleAddBook = () => {
-    navigate('/search');
-  };
-
-  const handleToggle = () => {
-    // TODO 디자인 변경시 수정 예정
-  };
-
-  return (
-    <>
-      <Header
-        leftBtn={{
-          icon: <HiPlus style={{ width: '24px', height: '24px' }} />,
-          handleLeftBtnClick: handleAddBook,
-        }}
-        rightBtn={{
-          icon: <BsThreeDotsVertical style={{ width: '24px', height: '24px' }} />,
-          handleRightBtnClick: handleToggle,
-        }}
-        title={{ text: '전체보기' }}
-      />
-      서재
-    </>
-  );
-};
-
-export default Library;
